Rename icon file and make close icon minimize to tray
The icon file for *odmon* was named inappropiately dock_icon.
It is now renamed to odmon_icon throughout the application.

Closing the *odmon* main window does not close the application
but rather minimizes to the system tray.  A 'Quit' button
is added to the GUI, so the *odmon* can be closed from the
main window.