Add some info to the drive tabs. Read the drive config.
We add the 'id', the z-base-32 encoding of the name and
the syncronization directory.

For the latter we have to read the `config` file in the
drive configuration directory.  Its content is made
available in the config($drive,config) array slot as a
dictionary.  Beware, you have to trim double quotes from
the values by yourself.